src/requirements.txt
#Pillow-SIMD>=6.0.0.post0
Pillow>=6.2.1
boto3==1.10.28